+++ core/helps.py
@@ -15,7 +15,6 @@ def print_help():
 		[ "%-25s" %("-p <path_to_wordlist>"), "-p /usr/share/wordlists/fasttrack.txt"],
 		[ "%-25s" %("-U <username>"), "-U admin | -U admin:user1"],
 		[ "%-25s" %("-t <threads>"), "-t 32"],
-		[ "%-25s" %("-k <false_key>"), "-k 'Invalid username'"]
 	]
 	utils.print_table(title, *menu)