+++ modules/httpget.py
@@ -21,13 +21,13 @@ def submit(URL, tryUsername, tryPassword, proxy, verbose, result):
 		
 		if resp.status_code == 200:
 			result.put([tryUsername, tryPassword])
-			utils.printf("[*] Match found: %s:%s" %(tryUsername, tryPassword), "good") 
+			utils.printf("[*] Match found: %ss" %([tryUsername, tryPassword]), "good") 
 			# pass
 		elif resp.status_code == 401:
 			if verbose:
-				utils.printf("[-] Failed: %s:%s" %(tryUsername, tryPassword), "bad")
+				utils.printf("[-] Failed: %ss" %([tryUsername, tryPassword]), "bad")
 		else:
 			# unknown
 			pass
 	except Exception as err:
-		utils.printf("[x] Error: %s:%s\n%s at %s" %(tryUsername, tryPassword, error, URL), "bad")
+		utils.printf("[x] Error: %s\n%s at %s" %([tryUsername, tryPassword], error, URL), "bad")